Create folders for components, create css file for each component
