+++ client/src/pages/AboutPage/components/GovernanceSection/GovernanceSection.jsx
@@ -1,3 +1,5 @@
+import "./GovernanceSection.css";
+
 const GovernanceSection = () => {
 	return (
 		<div className="governance-section">
