+++ client/src/pages/AboutPage/components/HistorySection/HistorySection.jsx
@@ -1,0 +1,28 @@
+import "./HistorySection.css";
+
+import image from "../../../assets/grey_image.svg";
+
+const HistorySection = () => {
+	return (
+		<div className="history-section">
+			<h1 className="history-title">History</h1>
+			<div className="title-line"></div>
+			<div className="history-wrapper">
+				<img src={image} alt="grey square" />
+				<p>
+					Lorem Ipsum is simply dummy text of the printing and typesetting
+					industry. Lorem Ipsum has been the industrys standard dummy text ever
+					since the 1500s, when an unknown printer took a galley of type and
+					scrambled it to make a type specimen book. It has survived not only
+					five centuries, but also the leap into electronic typesetting,
+					remaining essentially unchanged. It was popularised in the 1960s with
+					the release of Letraset sheets containing Lorem Ipsum passages, and
+					more recently with desktop publishing software like Aldus PageMaker
+					including versions of Lorem Ipsum.
+				</p>
+			</div>
+		</div>
+	);
+};
+
+export default HistorySection;
